+++ components/calendar/src/julian.rs
@@ -420,7 +420,7 @@ mod test {
             expected_year: i32,
             expected_era: Era,
             expected_month: u8,
-            expected_day: u32,
+            expected_day: u8,
         }
 
         let cases = [
